Address review feedback on PR #3314
- Add a wire format version tag to ExecutionWitness serialization: the
  first byte of every serialized witness is a version constant, and
  deserialization only accepts the current value, so future format
  changes have an explicit evolution point
- Drop source_node_ids from the serialized ContinuationStack: those
  indices reference a package's PackageDebugInfo, which is not on the
  wire, so they would be dangling on the deserialized side; a restored
  stack simply does not track source nodes, as a non-debug-aware
  execution would
- Document that Continuation serialization deliberately omits
  package_debug_info, so round-tripping any struct containing a
  Continuation is not exact, and note why the VecDeque serialization
  uses a newtype view and why ACE replay serialization uses free
  functions (the orphan rule forbids impls on the foreign types)
- Move the trace fragment state serialization impls into a
  trace_state/serde submodule to keep trace_state.rs focused
- Update the changelog phrasing away from the outdated trace proving
  input wording and add a README for the test-serde-macros helper crate
