+++ processor/src/continuation_stack.rs
@@ -176,14 +176,11 @@ impl<F> Continuation<F> {
 /// This allows the processor to execute a program iteratively in a loop rather than recursively
 /// traversing the nodes. It also allows the processor to pass the state of execution to another
 /// processor for further processing, which is useful for parallel execution of MAST forests.
-#[cfg_attr(
-    all(feature = "arbitrary", test),
-    miden_test_serde_macros::serde_test(
-        binary_serde(true),
-        serde_test(false),
-        types(MastForestId)
-    )
-)]
+///
+/// Note: the binary wire format for this type is deliberately lossy (`package_debug_info` and
+/// `source_node_ids` are not serialized), so the exact-equality round-trip test generated by
+/// `serde_test` does not apply; `continuation_stack_mast_forest_id_round_trip_omits_debug_metadata`
+/// covers the lossy round trip instead.
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct ContinuationStack<F> {
     stack: Vec<Continuation<F>>,
@@ -455,6 +452,11 @@ const TAG_RESPAN: u8 = 7;
 const TAG_FINISH_BASIC_BLOCK: u8 = 8;
 const TAG_ENTER_FOREST: u8 = 9;
 
+// NOTE: `package_debug_info` is deliberately *not* serialized: it is reconstructable from the
+// package a witness was produced from and would bloat the wire with data irrelevant to proving.
+// Round-tripping a `Continuation` (or any struct containing one, such as `ContinuationStack`)
+// therefore does not restore debug info exactly: deserialized continuations always carry
+// `package_debug_info: None`.
 impl Serializable for Continuation<MastForestId> {
     fn write_into<W: ByteWriter>(&self, target: &mut W) {
         match self {
@@ -545,27 +547,20 @@ fn read_mast_node_id<R: ByteReader>(source: &mut R) -> Result<MastNodeId, Deseri
     Ok(MastNodeId::from(u32::read_from(source)?))
 }
 
+// `source_node_ids` is deliberately *not* serialized: those indices point into a package's
+// `PackageDebugInfo`, which is itself absent from the wire (see the note above), so they would be
+// dangling references on the deserialized side. A restored stack simply does not track source
+// nodes, matching a non-debug-aware execution, which is all proving requires.
 impl Serializable for ContinuationStack<MastForestId> {
     fn write_into<W: ByteWriter>(&self, target: &mut W) {
         self.stack.write_into(target);
-        self.source_node_ids.write_into(target);
     }
 }
 
 impl Deserializable for ContinuationStack<MastForestId> {
     fn read_from<R: ByteReader>(source: &mut R) -> Result<Self, DeserializationError> {
         let stack = Vec::<Continuation<MastForestId>>::read_from(source)?;
-        let source_node_ids = Option::<Vec<Option<DebugSourceNodeId>>>::read_from(source)?;
-        if let Some(source_node_ids) = &source_node_ids
-            && source_node_ids.len() != stack.len()
-        {
-            return Err(DeserializationError::InvalidValue(format!(
-                "continuation source_node_ids length {} does not match stack length {}",
-                source_node_ids.len(),
-                stack.len()
-            )));
-        }
-        Ok(Self { stack, source_node_ids })
+        Ok(Self { stack, source_node_ids: None })
     }
 }
 
@@ -745,7 +740,7 @@ mod tests {
     }
 
     #[test]
-    fn continuation_stack_mast_forest_id_round_trip_omits_package_debug_info() {
+    fn continuation_stack_mast_forest_id_round_trip_omits_debug_metadata() {
         let mut stack: ContinuationStack<MastForestId> = ContinuationStack::default();
         stack.push_continuation(Continuation::StartNode(MastNodeId::from(1)));
         stack.push_continuation(Continuation::EnterForest {
@@ -785,9 +780,8 @@ mod tests {
                 op_idx_in_batch: 5,
             } if node_id == MastNodeId::from(3)
         ));
-        assert_eq!(
-            restored.source_node_ids,
-            Some(vec![Some(DebugSourceNodeId::from(10)), None, Some(DebugSourceNodeId::from(11)),])
-        );
+        // source_node_ids is not serialized: indices without their owning package debug info
+        // would be dangling on the restored side.
+        assert_eq!(restored.source_node_ids, None);
     }
 }
